+++ packages/client/src/widgets/unix-clock.vue
@@ -1,14 +1,12 @@
 <template>
-<div class="mkw-digitalClock _monospace" :class="{ _panel: !widgetProps.transparent }" :style="{ fontSize: `${widgetProps.fontSize}em` }">
+<div class="mkw-unixClock _monospace" :class="{ _panel: !widgetProps.transparent }" :style="{ fontSize: `${widgetProps.fontSize}em` }">
+	<div v-if="widgetProps.showLabel" class="label">UNIX time</div>
 	<div class="time">
-		<span v-text="hh"></span>
-		<span class="colon" :class="{ showColon }">:</span>
-		<span v-text="mm"></span>
-		<span class="colon" :class="{ showColon }">:</span>
 		<span v-text="ss"></span>
 		<span v-if="widgetProps.showMs" class="colon" :class="{ showColon }">:</span>
 		<span v-if="widgetProps.showMs" v-text="ms"></span>
 	</div>
+	<div v-if="widgetProps.showLabel" class="label">UTC</div>
 </div>
 </template>
 
@@ -17,7 +15,7 @@ import { onUnmounted, ref, watch } from 'vue';
 import { useWidgetPropsManager, Widget, WidgetComponentEmits, WidgetComponentExpose, WidgetComponentProps } from './widget';
 import { GetFormResultType } from '@/scripts/form';
 
-const name = 'digitalClock';
+const name = 'unixClock';
 
 const widgetPropsDef = {
 	transparent: {
@@ -30,6 +28,10 @@ const widgetPropsDef = {
 		step: 0.1,
 	},
 	showMs: {
+		type: 'boolean' as const,
+		default: true,
+	},
+	showLabel: {
 		type: 'boolean' as const,
 		default: true,
 	},
@@ -50,12 +52,10 @@ const { widgetProps, configure } = useWidgetPropsManager(name,
 );
 
 let intervalId;
-const hh = ref('');
-const mm = ref('');
 const ss = ref('');
 const ms = ref('');
 const showColon = ref(false);
-let prevSec: number | null = null;
+let prevSec: string | null = null;
 
 watch(showColon, (v) => {
 	if (v) {
@@ -67,12 +67,10 @@ watch(showColon, (v) => {
 
 const tick = () => {
 	const now = new Date();
-	hh.value = now.getHours().toString().padStart(2, '0');
-	mm.value = now.getMinutes().toString().padStart(2, '0');
-	ss.value = now.getSeconds().toString().padStart(2, '0');
-	ms.value = Math.floor(now.getMilliseconds() / 10).toString().padStart(2, '0');
-	if (now.getSeconds() !== prevSec) showColon.value = true;
-	prevSec = now.getSeconds();
+	ss.value = Math.floor(now.getTime() / 1000).toString();
+	ms.value = Math.floor(now.getTime() % 1000 / 10).toString().padStart(2, '0');
+	if (ss.value !== prevSec) showColon.value = true;
+	prevSec = ss.value;
 };
 
 tick();
@@ -94,9 +92,14 @@ defineExpose<WidgetComponentExpose>({
 </script>
 
 <style lang="scss" scoped>
-.mkw-digitalClock {
+.mkw-unixClock {
 	padding: 16px 0;
 	text-align: center;
+
+	> .label {
+		font-size: 65%;
+		opacity: 0.7;
+	}
 
 	> .time {
 		> .colon {