+++ CHANGELOG.md
@@ -14,6 +14,7 @@ You should also include the user name that made the change.
 
 ### Improvements
 - Client: Add vi-VN language support
+- Client: Add unix time widget @syuilo
 
 ### Bugfixes
 - Server: リモートユーザーを正しくブロックできるように修正する @xianonn